backend/pyproject.toml
[project]
name = "yunke-transit-backend"
version = "0.1.0"
description = "云客中转后端服务"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.0",
    "python-multipart>=0.0.6",
    "httpx[http2]>=0.27.0",
    "apscheduler>=3.10.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "ruff>=0.14.8",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

